+++ src/domain/index.ts
@@ -1,3 +1,4 @@
 export * from "./behavior.js";
 export * from "./defaults.js";
+export * from "./settings.js";
 export * from "./types.js";